Add tests that simulate uplugging devices on active streams
Create bunch of tests that can simulate uplugging the running devices on
the active streams. The device uplugged could be default device or not.
There is a time setting in those test to wait for the stream
reinitialization. If it's 0, then the stream is likely to be destroyed
before being reinitialized.