Simplify the APIs for querying device information (#8)
* Replace audiounit_get_default_datasource by get_device_source

* Prevent memory leak in current_device

Memory will leak when the macOS device only has input device. In that
case, the ownership of the CString for the input source will be
transferred and never be retaken. The changes make sure the
current_device call is always successful hence the ownership of the
memory can always be retaken by calling device_destroy.

* Replace audiounit_get_acceptable_latency_range by get_device_buffer_frame_size_range

Most code in audiounit_get_acceptable_latency_range is duplicate of the
get_device_buffer_frame_size_range.

* Create get_device_latency

* Create get_device_streams

* Create get_stream_latency

* Create get_device_sample_rate

With the new introduced API, the nominal sample rate of the device can
be queried wihtout using audiounit_get_available_samplerate, which gives
additional information but it's not often useful.

* Replace audiounit_get_available_samplerate by get_range_of_sample_rates

* Replace audiounit_get_device_presentation_latency by get_presentation_latency

Add logs when the API calls fail before returning the default latency

* Move APIs for AudioDeviceID together

* Create get_device_stream_format

* Simplify preferred_sample_rate

* Simplify get_channel_count

* Shorten the variable name