Avoid various racing issues  (#26)
* Start AudioUnit in serial queue

The stream-start and stream-reinit may run in parallel and cause a
racing issue: The stream data callback may still run while the stream is
being reinitialized. Running stream-start in the same task queue where
the reinit task is could solve this racing issue.

* Stop AudioUnit in serial queue

The stream-stop and stream-reinit may run in parallel and causes a
racing issue: The stream-stop may executed before stream-restart in
reinit so the stream-stop fails silently. Running stream-stop in the
same task queue where reinit task is could solve this racing issue.

* Stop data-callback in serial queue when destorying stream

The stream-start and stream-destroy may run in parallel and cause a
racing issue: The data-callback may still be run while the stream is
being destroyed. The stream-stop in destroy fails silently in this case.
Moving stream-stop in the task queue where the stream-destroy and
stream-start are, beforing doing stream destroy task, could solve this
racing issue.

* Assert the timing for start AudioUnit

An an assertion to make sure the stream start is only allowed to be
called after the stream is initialized and before the stream is
destroyed.

* Rename variables

* Remove unnecessary mutex

The clone of the mutex that will be passed into sync_dispatch is not
really needed. Those data in the mutexes won't be operated in the same
time since the code after sync_dispatch won't be executed before the task
dispatched by sync_dispatched are finished. Hence, mutex around the data
that will be operated by the code in sync_dispatched can be omitted.

* Log state in state callback