Save the duplicate tests and compilings when running sanitizers (#52)
* Fix typo

* Remove the duplicate sanitizer tests

Since the tests in the sub crates will be run in run_tests.sh and the
run_tests.sh will be executed by the sanitizers, the tests in the sub
crates will be executed by the sanitizers with the command
`RUSTFLAGS="-Z sanitizer=*" sh run_tests.sh`. That is, running the tests
in under sub crate by `RUSTFLAGS="-Z sanitizer=*" cargo test` is
duplicate work.

* Simplify run_sanitizers.sh

* Reuse the compiling when building the sub crate

The run_tests.sh should use `cargo * -p <SUB_CRATE>` if it's possible.
By doing so, the compiling for <SUB_CRATE> can be reused when building
the main crate. It will save plenty of compiling time, especially when
compiling this crate for the sanitizers.