Implement a API to get `groupd_id` with special case handling (#69)
* Add test for get_device_model_uid

* skip checking group_id in test_create_cubeb_device_info

* Print error code when get_device_source* fails

* Remove dbg in get_device_stream_configuration

* new API: get_device_global_model_uid

* Assign group_id by get_device_global_model_uid

Enable the group_id check in `check_device_info_by_device` as well.

* Revert "Assign group_id by get_device_global_model_uid"

This reverts commit 8359f4592b870a586693575cb4bd021b7cfeeff6.

* Revert "new API: get_device_global_model_uid"

This reverts commit f49b681881143058c54847ad0310678f476c5978.

* Sort `Property` in alphabetical order

* Add an API to get the transport type

The transport type will be used to check the device is built-in or not.

* clearer log for test_get_device_source

* Add an API to get `group_id` with special cases

This API will return same group_ip for <internal mic, internal speaker>
and <TRRS mic, TRRS headphone> if the devices are builtin. Otherwise, it
returns the model uid instead. This should address #73.

* Get model uid in global scope when querying in input or output fails

* Move get_device_label to mod.rs

`get_device_label` is not really a low-level API, so move it from
device_property.rs to mod.rs.

* Move get_device_global_uid to mod.rs

`get_device_global_uid` is not really a low-level API, so it should be
moved from device_property.rs to mod.rs

The APIs in device_property.rs should map to what `Property` has.

* Return `CString` instead of `StringRef` from `get_device_group_id`

`StringRef::from_str` can be removed by this change.

* Test builtin device pairs have same group_id

* Assert  fails for unknown device