Return an error if sub-device-change cannot be monitored (#109)
This patch tries to address [BMO 1658982](https://bugzilla.mozilla.org/show_bug.cgi?id=1658982).

One of the assertions in `set_sub_devices_sync` can be hit in the wild.
The crash-stack doesn't show which assertion is hit. My guess is that
the newly created aggregate-device may not be ready yet, to be used to
watch its own property-change, even if the created device is already in
the process-wide device list.