BMO 1721496 - Append silence frames to when switching from A2DP to HFP/HSP (#132)
* BMO 1721496 - Append silence frames to when switching from A2DP to HFP/HSP

No silence frame is fed to resampler when the duplex stream runs on a
bluetooth when the device is switchiing from A2DP to HFP/HSP profile,
which leads to hit an assertion [1] within cubeb resampler.

When the duplex cubeb stream runs on a bluetooth device and the device
is switching from A2DP to HFP/HSP profile, the `input_buffer_manager`
has no input data to be appended [2] in `audiounit_input_callback`. And
then when `audiounit_input_callback` runs, the `buffered_input_frames`
is `0` but `input_frames_needed` is a positive integer greater than `0`.
However, the `input_frames` in this case is equal to
`buffered_input_frames` since neither `switching_device` nor
`frames_read > 0` is `true`.

When the above case happens in the first `audiounit_output_callback`
call, it makes the cubeb resampler hit the assertion in [1]. The `data_`
[3] in the `internal_input_buffer` [4] is `nullptr` when the resampler
is created (`passthrough_resampler` in my case). If the `length` in the
first `push*` [5] is `0`, then the `reserve` won't be called so `data_`
remains `nullptr`, which makes the resampler hit the assertion in [1]

We should append the silence frames to the resampler in this case not
only to avoid hitting the assertion but also to append the silence
frames correctly when we need.

[1] https://github.com/mozilla/cubeb/blob/b608f590249d5821c91b3ea83e394f0fd839f5b9/src/cubeb_utils.h#L31
[2] https://github.com/mozilla/cubeb-coreaudio-rs/blob/40b3baf3ba909201697702418c4a74f8a05f2ec3/src/backend/mod.rs#L409
[3] https://github.com/mozilla/cubeb/blob/b608f590249d5821c91b3ea83e394f0fd839f5b9/src/cubeb_utils.h#L115
[4] https://github.com/mozilla/cubeb/blob/b608f590249d5821c91b3ea83e394f0fd839f5b9/src/cubeb_resampler_internal.h#L112
[5] https://github.com/mozilla/cubeb/blob/b608f590249d5821c91b3ea83e394f0fd839f5b9/src/cubeb_utils.h#L196
[6] https://github.com/mozilla/cubeb/blob/b608f590249d5821c91b3ea83e394f0fd839f5b9/src/cubeb_utils.h#L172

* Remove unnecessary checks

* Correct frames_read

`frames_read` is added in `audiounit_input_callback` with K frames but
it can be added by K+N frames again when `input_frame_needed` in
`audiounit_output_callback is K+N, where K is non-negative integer and N
is positive integer. The N is the number of the silence frames we add in
the input buffer when input_buffer_manager doesn't have enough frames
resampler needs. The `frames_read` should be added by N, which is
silence frames only, rather than K+N in this case.

* Make `frames_read` always follow input_buffer_manager

We always use `input_buffer_manager.get_linear_data(...)` to get the
input buffer with required frames, so `frames_read` should be the sum of
the required frames.