Exclude CubebAggregateDevice when getting device list
We might create a custom cubeb-private aggregate device as our
input-output device when creating a duplex stream. The lifetime of this
aggregate device follows the lifetime of the created duplex stream.

We should not expose it in our test-only device-list getter API.
Otherwise, if we set it to another stream's input or output device, that
aggregate device may die before we finish the test for another stream.

We've done the same thing in our cubeb API. We should do the same for
the test-only API as well.