Don't switch input when using non-default device
By the comments [1,2] in the BMO 1735201, we should not switch the input
device behind the scenes when users specify what devices they want. That
is, we should not automatically re-initialize the input or duplex stream
when the input device is gone. We will get a device-collection changed
event as well when this happen and Gecko/Firefox will stop the paired
MediaStreamTrack soon in this case.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1735201#c0
[2] https://bugzilla.mozilla.org/show_bug.cgi?id=1735201#c11